This is the working game
This is the working game, it slowly gets faster and all of the components work. I still need to add the images, and the intro and finishing screens. As I have figured out the speed change, there is no need to have a game with only them, so I removed them from the hub